feat(order): 使用 BigDecimal 替代 Double以提高精度
- 将 OrderItemEntity 中的 subtotal 类型从 Double 改为 BigDecimal
- 在 OrderServiceImpl 中使用 BigDecimal 进行金额计算
- 修正商品总价计算公式，确保精度不丢失
- 更新订单明细小计金额的计算方式- 避免浮点数运算带来的精度问题

+++ common/src/main/java/com/ligg/common/module/entity/OrderItemEntity.java
@@ -11,6 +11,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 /**
  * 订单明细
  */
@@ -50,5 +52,5 @@ public class OrderItemEntity {
     /**
      * 小计金额(价格x数量)
      */
-    private Double subtotal;
+    private BigDecimal subtotal;
 }
